+++ src/main/java/com/modular/restfulserver/user/dto/UserInfo.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 
 @Getter
-public class UserDetails {
+public class UserDetail {
 
   private final Long id;
   private final String email;
@@ -20,7 +20,7 @@ public class UserDetails {
   private final LocalDateTime updatedDate;
 
   @Builder(setterPrefix = "add")
-  public UserDetails(Long id, String email, String username, String realname, LocalDateTime createdDate, LocalDateTime updatedDate) {
+  public UserDetail(Long id, String email, String username, String realname, LocalDateTime createdDate, LocalDateTime updatedDate) {
     Assert.isInstanceOf(Long.class, id, getIllegalFieldError("userId"));
     Assert.isInstanceOf(String.class, email, getIllegalFieldError("email"));
     Assert.isInstanceOf(String.class, username, getIllegalFieldError("username"));
@@ -36,8 +36,8 @@ public class UserDetails {
     this.updatedDate = updatedDate;
   }
 
-  public static UserDetails from(User user) {
-    return UserDetails.builder()
+  public static UserDetail from(User user) {
+    return UserDetail.builder()
       .addId(user.getId())
       .addEmail(user.getEmail())
       .addUsername(user.getUsername())
