+++ src/test/java/com/modular/restfulserver/user/dto/UserDetailTest.java
